# swagger.yaml
openapi: 3.0.3
info:
  title: Crypto Trading Platform API
  description: |
    Complete API documentation for the crypto trading platform.
    Includes REST APIs and WebSocket connections for real-time data.
    
    Base URL: http://192.168.1.47:8080
    WebSocket URL: ws://192.168.1.47:8081/ws
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://192.168.1.47:8080
    description: Backend API Server
  - url: ws://192.168.1.47:8081
    description: WebSocket Server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or email verification
  schemas:
...
    VerifyEmailRequest:
      type: object
      required:
        - uuid
        - code
      properties:
        uuid:
          type: string
          format: uuid
          description: UUID received from signup
        code:
          type: string
          description: Verification code sent to email
          example: "123456"
    VerifyEmailResponse:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
          description: JWT authentication token
        message:
          type: string
        status:
          type: integer
...
    Holding:
      type: object
      properties:
        id:
          type: integer
        chain_id:
          type: integer
        token_address:
          type: string
        type:
          type: integer
          description: Transaction type
        amount_native:
          type: number
          description: Amount in native currency (lamports)
        amount_token:
          type: number
          description: Amount in tokens
        created_at:
          type: integer
          description: Unix timestamp
        balance:
          type: string
        updated_at:
          type: integer
        name:
          type: string
        symbol:
          type: string
        meta_data:
          type: string
    HoldingsResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        holdings:
          type: array
          items:
            $ref: '#/components/schemas/Holding'
        total:
          type: integer
        pagination:
          $ref: '#/components/schemas/Pagination'
    Transaction:
      type: object
      properties:
        id:
          type: integer
        chain_id:
          type: integer
        token_address:
          type: string
        tx_hash:
          type: string
        type:
          type: integer
          description: 0 = Sell, 1 = Buy
          enum: [0, 1]
        amount_native:
          type: number
        amount_token:
          type: number
        created_at:
          type: integer
          description: Unix timestamp
        name:
          type: string
        symbol:
          type: string
        meta_data:
          type: string
    TransactionsHistoryResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        total:
          type: integer
        pagination:
          $ref: '#/components/schemas/Pagination'
    Position:
      type: object
      properties:
        id:
          type: integer
        chain_id:
          type: integer
        token_address:
          type: string
        tx_hash:
          type: string
        type:
          type: integer
          description: 0 = Sell, 1 = Buy
          enum: [0, 1]
        amount_native:
          type: number
        amount_token:
          type: number
        created_at:
          type: integer
    PositionsResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        positions:
          type: array
          items:
            $ref: '#/components/schemas/Position'
        total:
          type: integer
        pagination:
          $ref: '#/components/schemas/Pagination'
    Holder:
      type: object
      properties:
        wallet_address:
          type: string
        balance:
          type: string
        percent:
          type: number
          description: Percentage of total supply
    HoldersResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        holders:
          type: array
          items:
            $ref: '#/components/schemas/Holder'
        total:
          type: integer
        pagination:
          $ref: '#/components/schemas/Pagination'
    Pagination:
      type: object
      properties:
        offset:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        has_more:
          type: boolean
          
    # WebSocket Schemas
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        message:
          type: string
paths:
...
  /api/user/verify_email:
    post:
      tags:
        - Authentication
      summary: Verify email address
      description: Verify user email with code sent to email. Returns JWT token on success.
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: Email verified successfully
          headers:
            Set-Cookie:
              schema:
                type: string
                example: auth_token=eyJhbGc...; Path=/; HttpOnly; SameSite=Lax
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyEmailResponse'
        '400':
          description: Invalid verification code or UUID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/tokens/holding:
    get:
      tags:
        - Tokens
      summary: Get user holdings
      description: Retrieve user's token holdings with pagination
      operationId: getUserHoldings
      security:
        - BearerAuth: []
      parameters:
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Pagination offset
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
          description: Number of items per page
      responses:
        '200':
          description: Holdings retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HoldingsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tokens/history:
    get:
      tags:
        - Tokens
      summary: Get transaction history
      description: Retrieve user's transaction history with optional filters
      operationId: getTransactionHistory
      security:
        - BearerAuth: []
      parameters:
        - name: chainId
          in: query
          schema:
            type: integer
          description: Filter by chain ID
        - name: tokenAddress
          in: query
          schema:
            type: string
          description: Filter by token address
        - name: type
          in: query
          schema:
            type: integer
            enum: [0, 1]
          description: Filter by transaction type (0=Sell, 1=Buy)
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Pagination offset
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
          description: Number of items per page
      responses:
        '200':
          description: Transaction history retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionsHistoryResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
                
  # Token Detail Endpoints
  /api/tokens/{chainId}/{tokenAddress}/positions:
    get:
      tags:
        - Token Details
      summary: Get token positions
      description: Retrieve all positions (trades) for a specific token
      operationId: getTokenPositions
      security:
        - BearerAuth: []
      parameters:
        - name: chainId
          in: path
          required: true
          schema:
            type: integer
            example: 501
          description: Blockchain chain ID
        - name: tokenAddress
          in: path
          required: true
          schema:
            type: string
          description: Token contract address
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Pagination offset
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
          description: Number of items per page
      responses:
        '200':
          description: Positions retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PositionsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tokens/{chainId}/{tokenAddress}/holders:
    get:
      tags:
        - Token Details
      summary: Get token holders
      description: Retrieve list of token holders with their balances
      operationId: getTokenHolders
      security:
        - BearerAuth: []
      parameters:
        - name: chainId
          in: path
          required: true
          schema:
            type: integer
            example: 501
          description: Blockchain chain ID
        - name: tokenAddress
          in: path
          required: true
          schema:
            type: string
          description: Token contract address
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Pagination offset
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
          description: Number of items per page
      responses:
        '200':
          description: Holders retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HoldersResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/metadata:
    get:
      tags:
        - Tokens
      summary: Get token metadata
      description: Proxy endpoint to fetch token metadata from external URL
      operationId: getTokenMetadata
      parameters:
        - name: url
          in: query
          required: true
          schema:
            type: string
            format: uri
          description: URL to fetch metadata from
      responses:
        '200':
          description: Metadata retrieved successfully
          content:
            application/json:
              schema:
                type: object
                description: Token metadata (structure varies by source)
        '400':
          description: URL parameter is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to fetch metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

# WebSocket Documentation
...